Clean up create-react-app default boilerplate
I decided to keep most of the default boilerplate because some of
it is currently useful:

- React.StrictMode in index.js
  - good for development and logging things to console
- index.css
  - even though I want to use styled components for basically all styling
    throughout the app, I want to keep this because it's good for applying
    global styles
  - the `body` css boilerplate looks solid
- robots.txt
  - good for telling web crawlers like Googlebot what to look at

and other things might be useful later on:

- if I want to add unit tests
  - setupTests.js
  - App.test.js
- if I want to make this a PWA later
  - manifest.json (and the associated logo.png's)
  - the rel="apple-touch-icon" link tag in index.html
  - serviceWorker.js

So the only real changes here are:

- Refactor App.js to be really basic and use styled components
  - This means removing App.css and logo.svg
- Remove the `code` selector css from index.css
  - I don't anticipate having any `code` HTML tags in my jsx

I anticipate making changes to the index.html (like the favicon) and
README later on once I get closer to having people actually looking
at my project. But for now I'll just leave the boilerplate.