Address 'react-hooks/exhaustive-deps' warning
Not sure why this warning is being triggered since
adjustContainerSizeAndTreePosition doesn't reference any
props nor state.

https://reactjs.org/docs/hooks-faq.html#is-it-safe-to-omit-functions-from-the-list-of-dependencies

In any case, one way to address this warning is to define the
function within the useEffect. We won't do that though because
we call the function in another place as well.

So instead we pass the function as a dependency to useEffect. We
wrap it in useCallback so we don't get a new function reference
everytime the component renders (which is what React normally
does) thereby unnecessarily triggering the useEffect.