Add custom pathFunc
This replaces the need for hidePathsToBlankNodes. This is
amazing because we no longer have to worry about errors where
hidePathsToBlankNodes sometimes gets called before the tree
is rendered (and therefore the paths aren't available in the DOM
yet). The unneeded paths are already 'dealt with' from the get-go
with pathFunc, instead of retroactively with hidePathsToBlankNodes.

This also allows the tree g element (and therefore tree container)
to not be unnecessarily wide in case there were unneeded paths
at the edges leading to blank nodes. Now the unneeded paths are gone
so they're no longer accounted for in react-d3-tree calculating the
tree g element's width. This matters for Mobile.

It looks like react-d3-tree doesn't account for the nodes themselves
in calculating the width of the tree g element so we don't need to
deal with the blank nodes that might be on the edges.

For more context on the usage of pathFunc, see
https://github.com/bkrem/react-d3-tree/issues/214.